pyproject.toml
[project]
name = "mads-siads-699-winter-2026-capstone"
version = "0.1.0"
description = "MADS SIADS 699: Capstone Project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=5.5.0",
    "bokeh>=3.8.0",
    "colorcet>=3.1.0",
    "datashader>=0.18.2",
    "holoviews>=1.21.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "polars>=1.33.1",
    "polars-lts-cpu>=1.33.1",
    "pyarrow>=21.0.0",
    "pydantic-settings>=2.10.1",
    "pygithub>=2.8.1",
    "pygraphviz>=1.14",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "statsmodels>=0.14.5",
    "umap-learn>=0.5.9.post2",
    "vegafusion>=2.0.2",
    "vl-convert-python>=1.8.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "jupyterlab>=4.4.7",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-ruff>=0.5",
    "ruff>=0.13.0",
]

[tool.ruff]
extend-include = ["*.ipynb"]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".ipynb_checkpoints",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist"
]

line-length = 120
indent-width = 4
target-version = "py313"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
